Refactor code to use custom hook for data loading
