+++ src/App.tsx
@@ -8,43 +8,17 @@ import "./css/input.css";
 import { RouterProvider, createBrowserRouter } from "react-router-dom";
 import { useAppContext } from "./store/app.context";
 import { useState, useEffect } from "react";
-import data from "./data.json";
 import { routes } from "./config";
 import EmployeeDeletePopup from "./pages/EmployeeListing/components/EmployeeDeletePopup";
+import useLoadData from "./hooks/useLoadData";
 
 const router = createBrowserRouter(routes);
 
 function App() {
-  const [loading, setLoading] = useState(true);
-  const appContext = useAppContext();
-
-  const loadData = async () => {
-    let employees =
-      localStorage.getItem("employees") &&
-      JSON.parse(localStorage.getItem("employees") as string);
-    let skills =
-      localStorage.getItem("skills") &&
-      JSON.parse(localStorage.getItem("skills") as string);
-    let departments =
-      localStorage.getItem("departments") &&
-      JSON.parse(localStorage.getItem("departments") as string);
-    if (skills === null || employees === null || departments === null) {
-      employees = data.employees;
-      skills = data.skills;
-      departments = data.departments;
-    }
-    appContext.dispatch({ type: "SET_EMPLOYEES", payload: employees });
-    appContext.dispatch({ type: "SET_SKILLS", payload: skills });
-    appContext.dispatch({ type: "SET_DEPARTMENTS", payload: departments });
-    setLoading(false);
-  };
-
-  useEffect(() => {
-    loadData();
-  }, []);
+  const loading = useLoadData();
 
   if (loading) {
-    return <div>Loading1...</div>;
+    return <div>Loading...</div>;
   }
 
   return (
