+++ src/hooks/useQuery.ts
@@ -2,7 +2,7 @@ import React from "react";
 import { useLocation } from "react-router-dom";
 
 export default function useQuery() {
-    const { search } = useLocation();
-  
-    return React.useMemo(() => new URLSearchParams(search), [search]);
-  }
+  const { search } = useLocation();
+
+  return React.useMemo(() => new URLSearchParams(search), [search]);
+}
